backend/requirements.txt
# Core framework
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv

# LangChain ecosystem
langchain
langchain-core
langchain-openai
langchain-community
langgraph

# Vector DB
pinecone

# HuggingFace
huggingface_hub

# Auth
pyjwt
bcrypt

# Utilities
httpx
